requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.4
reportlab==4.0.7
Werkzeug==2.3.7
gunicorn==21.2.0
itsdangerous==2.1.2
click==8.1.7
Jinja2==3.1.2
qrcode[pil]==7.4.2
Pillow==10.2.0
openpyxl==3.1.2
requests==2.31.0
plotly==5.18.0
psycopg2-binary==2.9.9
schedule==1.2.1
opencv-python==4.9.0.80

# AI / ML additions
tensorflow-cpu==2.15.0
